Make macOS-specific alacritty setup only happen on macOS
Seems like the configuration that was there for macOS was invalid for my
Ubuntu setup. I was hoping it would just not work when not present for
the OS. And not sure there is a way to conditionally set it in YAML, so
copying the way I did it in the tmux setup.