Update tmux config
Move tmux into ~/.config/tmux directory.

Also some changes to the actual config, the nicest thing was to add:

set -sg escape-time 0

Which makes vim a bit faster, since we are not waiting for escape when
doing visual block, etc. However, it does make it so that I cannot use
Meta then ESC-2 to make the splits equal in size, need to figure out
something for that, but I appreciate this better than the alternative
for now.